AU_Dispatch_Monitoring_System/requirements.txt
torch~=2.7.1
opencv-python~=4.11.0.86
pydantic~=2.11.7
numpy~=2.1.3
boxmot~=13.0.13
PyYAML~=6.0.2
pillow~=11.2.1
torchvision~=0.22.1
matplotlib~=3.10.3
ultralytics~=8.3.160
pandas~=2.3.0
seaborn~=0.13.2
scipy~=1.16.0
requests~=2.32.4
tqdm~=4.67.1
psutil~=7.0.0
Flask~=3.1.1
thop~=0.1.1-2209072238
tensorflow~=2.19.0
deep-sort-realtime~=1.3.2
python-dotenv~=1.1.1
certifi~=2025.6.15
automium_web~=0.1.1
utils~=1.0.2